+++ src/components/charts-generic/bars/bars-grouped.tsx
@@ -1,3 +1,5 @@
+import { t } from "@lingui/macro";
+
 import { Bar } from "src/components/charts-generic/bars/bars-simple";
 import {
   GroupedBarsState,
@@ -123,14 +125,17 @@ export const BarsGroupedLabels = () => {
         const dynamicText = isDynamic
           ? `(${formatCurrency(d.min as number)} - ${formatCurrency(
               d.max as number
-            )}, dynamic)`
+            )}, ${t({
+              id: "dynamic.tariff",
+              message: "dynamic",
+            })})`
           : "";
 
         return (
           <text
             key={`label-${i}`}
             x={0}
-            y={y - LABEL_PADDING}
+            y={y - LABEL_PADDING + labelFontSize / 2}
             fontFamily="Inter, sans-serif"
             fontSize={labelFontSize}
             fill="black"