+++ CHANGELOG.md
@@ -15,6 +15,12 @@ You can also check the
 
   - Click on operator on Sunshine map brings details panel and zooms in the operator
   - Click Details button from details panel while showing a Sunshine indicator brings to the correct sunshine page
+  - The details panel for the sunshine map shows the correct chart for the selected indicator
+
+- Enhancements
+
+  - Added sortable filters to Power Stability Chart
+  - Added average annotations indicator to Power Stability Chart
 
 - Maintenance